[nextest-runner] move metadata reporting out to a separate file
We're also going to add JSON metadata output as well.
